chore: add flutter_lints
